refactor: 테스트 이름 변경
